Fixed font weights on titles for cards + moved priceRangeButton component to components dir
