+++ src/screens/RestaurantPicker/DietaryPref.js
@@ -1,9 +1,9 @@
-import React, {useState} from 'react'
+import React from 'react'
 import {StyleSheet, View,} from "react-native";
 import RestaurantCard from "../../components/RestaurantCard";
 import RestaurantQuestion from "../../components/RestaurantQuestion";
 import TagsSection from '../../components/TagsSection';
-import SmallText from '../../components/SmallText';
+import MediumText from '../../components/MediumText';
 
 const DietaryPref = ({setSelectedDietaryTags, selectedDietaryTags}) => {
 
@@ -15,7 +15,7 @@ const DietaryPref = ({setSelectedDietaryTags, selectedDietaryTags}) => {
                 <RestaurantQuestion text = {"Do you have any dietary preferences?"}/>
             </View>
             <View style = {styles.exampleTextContainer}>
-                <SmallText size = "13" color = "#A9A9A9" center={true}>E.g. Vegetarian, Gluten-free</SmallText>
+                <MediumText size = "13" color = "#A9A9A9" center={true}>E.g. Vegetarian, Gluten-free</MediumText>
             </View>
             <View style = {styles.textInputContainer}>
               <TagsSection
